+++ src/components/Loading/index.js
@@ -2,8 +2,10 @@ import React from "react";
 
 export default class Loading extends React.Component {
   render() {
-    return (<div className = "container">
-      <h3>Loading</h3>
-    </div>)
+    return (
+      <div className="container hv-center">
+        <h3>Loading...</h3>
+      </div>
+    );
   }
 }