+++ src/scenes/Share/index.js
@@ -1,5 +1,9 @@
 import React from "react";
 
+import FormInput from "components/FormInput";
+
+import "./styles.css";
+
 export default class Share extends React.Component {
   constructor() {
     super();
@@ -8,14 +12,58 @@ export default class Share extends React.Component {
 
   componentDidMount() {
     // Change page title
-    document.title = "New New | Share"
+    document.title = "New New | Share";
   }
 
   render() {
     return (
-      <div className="container">
-        <h5>Share</h5>
+      <div className="container hv-center">
+        <div className="row justify-content-center">
+          <div className="col-12 col-md-5 col-lg-4 hv-center">
+            <div>
+              <div className="pb-3">
+                <h3>Know any artists?</h3>
+                <h5>
+                  Plug them in.
+                  <span className="ec ec-electric-plug" />
+                </h5>
+              </div>
+              <div>
+                <p>
+                  <h6>Sharing music is for the soul.</h6>
+                </p>
+                <p>
+                  From one listener to another, let's share music. Drop your
+                  current favourite artsit and your favourite work from them,
+                  and I'll add it to the New New collection. Also, don't forget
+                  to give me your name so I can shout you out.
+                </p>
+                <p>Love.</p>
+              </div>
+            </div>
+          </div>
+          <div className="col-12 col-md-5 col-lg-4 hv-center">
+            <div>
+              <div className="card">
+                <div className="card-body">
+                  <FormInput inputID="shareName" inputPlaceholder="Your Name" />
+                  <FormInput
+                    inputID="shareArtist"
+                    inputPlaceholder="Artist Name"
+                  />
+                  <FormInput
+                    inputID="shareNotable"
+                    inputPlaceholder="Notable Work"
+                  />
+                  <button type="submit" className="btn btn-primary">
+                    Submit
+                  </button>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
       </div>
-    )
+    );
   }
 }